24/2

+++ src/pages/Buses/BusesBuses.jsx
@@ -5,6 +5,8 @@ import NavBuses from './NavBuses.jsx'
 import ThreeThing from '../../component/ThreeThing.jsx';
 import delet from  '../../assets/delete.svg';
 import pin from '../../assets/pin.svg';
+import Swal from 'sweetalert2';
+
 const BusesBuses = () => {
     const [data, setData] = useState([]);
     const [update, setUpdate] = useState(false);
@@ -29,24 +31,39 @@ const BusesBuses = () => {
         });
     }, [update])
   
-     
-  const handleDelete = (index) => {
-    const token = localStorage.getItem('token');
-
-    axios.delete(`https://bcknd.ticket-hub.net/api/admin/bus/delete/${index}`, {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    })
-      .then(response => {
-        console.log('User deleted successfully:', response.data);
-        setUpdate(!update);
-
-      })
-      .catch(error => {
-        console.error('Error deleting user:', error);
+    const handleDelete = (index, userName) => { 
+      const token = localStorage.getItem('token');
+      
+      Swal.fire({
+        title: `Are you sure you want to delete  bus${userName}?`, 
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'Yes',
+        cancelButtonText: 'No',
+      }).then((result) => {
+        if (result.isConfirmed) {
+          axios.delete(`https://bcknd.ticket-hub.net/api/admin/bus/delete/${index}`, {
+            headers: {
+                Authorization: `Bearer ${token}`,
+              },
+            })
+            .then((response) => {
+              console.log('User deleted successfully:', response.data);
+              setUpdate(!update);
+              Swal.fire('Deleted!', `${userName} has been deleted successfully.`, 'success'); 
+            })
+            .catch((error) => {
+              console.error('Error deleting user:', error);
+              Swal.fire('Error!', `There was an error while deleting ${userName}.`, 'error'); 
+            });
+        } else {
+      
+          Swal.fire('Cancelled', `${userName} was not deleted.`, 'info');  
+        }
       });
-  }
+    };
+    
+  
 
   const handleEdit = (index) => {
     const snedData = data.find((item) => item.id === index);
@@ -89,7 +106,7 @@ const BusesBuses = () => {
                                      <img
                                        className='w-[24px] h-[24px] ml-2 cursor-pointer'
                                        src={delet}
-                                       onClick={() => handleDelete(item.id)}
+                                       onClick={() => handleDelete(item.id,item.bus_number)}   
                                        alt="delete"
                                      />
                     </td>
